(#21) Throw error if recovery_message contains '
Previously, if you declared an `osx::recovery_message` resource that
contained an apostrophe/single-quote ('), the exec used to set the
recovery message would fail (because the exec references the message
within single quotes).

This commit will call the fail() function if a recovery_message contains
an apostrophe and alert the user to the failure.  The spec was also updated
and worked for me in 1.8.7 and 1.9.3